*refactoring of ConvertScaledDotProductAttentionNode.
