+++ src/Cvt/OnnxRuntime/Convert.h
@@ -175,6 +175,8 @@ namespace Synet
 
     bool ConvertRoundNode(const onnx::NodeProto& node, LayerParam& layer);
 
+    bool ConvertScaledDotProductAttentionNode(const onnx::NodeProto& node, const LayerParams& layers, LayerParam& layer, Bytes& reordered);
+
     bool ConvertSignNode(const onnx::NodeProto& node, LayerParam& layer);
 
     bool ConvertSqueezeNode(const onnx::NodeProto& node, const LayerParams& layers, LayerParam& layer);
